Rename WindowsString to WideCString
To be more in line with convention of owned `(C)String` types
