+++ src/asymmetric.rs
@@ -9,7 +9,7 @@
 
 use crate::helpers::blob::Blob;
 use crate::helpers::blob::BlobLayout;
-use crate::helpers::{AlgoHandle, Handle, WindowsString};
+use crate::helpers::{AlgoHandle, Handle, WideCString};
 use crate::key::ErasedKeyBlob;
 use crate::key::{BlobType, KeyHandle};
 use crate::property::{AlgorithmName, EccCurveName};
@@ -144,7 +144,7 @@ impl AsymmetricAlgorithm {
             | AsymmetricAlgorithmId::Ecdsa(NamedCurve::NistP384)
             | AsymmetricAlgorithmId::Ecdsa(NamedCurve::NistP521) => {}
             AsymmetricAlgorithmId::Ecdh(curve) | AsymmetricAlgorithmId::Ecdsa(curve) => {
-                let property = WindowsString::from_str(curve.as_str());
+                let property = WideCString::from_str(curve.as_str());
 
                 handle.set_property::<EccCurveName>(property.as_slice_with_nul())?;
             }
@@ -165,7 +165,7 @@ impl AsymmetricAlgorithm {
         let name = self
             .handle
             .get_property_unsized::<AlgorithmName>()
-            .map(|name| WindowsString::from_bytes_with_nul(name).to_string())?;
+            .map(|name| WideCString::from_bytes_with_nul(name).to_string())?;
 
         AsymmetricAlgorithmId::try_from(name.as_str()).map_err(|_| crate::Error::InvalidHandle)
     }
